refactor: changed planner resource to plan
