+++ shared/src/commonMain/kotlin/edu/usf/cutr/otp/plan/api/PlanApi.kt
@@ -1,8 +1,8 @@
-package edu.usf.cutr.otp.planner.api
+package edu.usf.cutr.otp.plan.api
 
 import edu.usf.cutr.otp.ApplicationDispatcher
-import edu.usf.cutr.otp.planner.model.Planner
-import edu.usf.cutr.otp.planner.model.RequestParameters
+import edu.usf.cutr.otp.plan.model.Planner
+import edu.usf.cutr.otp.plan.model.RequestParameters
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import io.ktor.client.*
@@ -16,7 +16,7 @@ class PlannerApi(private val host: String,
 
     //private val testUrl = "http://10.0.2.2:8080/otp/routers/default/plan?fromPlace=41.84712%2c-87.64678&toPlace=41.84584%2c-87.65214"
 
-    fun getPlannerResource(
+    fun getPlan(
         success: (Planner) -> Unit, failure: (Throwable?) -> Unit) {
 
         GlobalScope.launch(ApplicationDispatcher) {
