+++ shared/src/commonMain/kotlin/edu/usf/cutr/otp/plan/model/Details.kt
@@ -1,14 +1,10 @@
-package edu.usf.cutr.otp.planner.model
+package edu.usf.cutr.otp.plan.model
 
 import kotlinx.serialization.Polymorphic
 import kotlinx.serialization.Serializable
 
-
 @Serializable
-data class Metadata
+data class Details
 (
-    val searchWindowUsed: Int? = null,
-    val nextDateTime: Long? = null,
-    val prevDateTime: Long? = null,
     val additionalProperties: MutableMap<String,@Polymorphic Any> = HashMap()
 )
