+++ shared/src/commonMain/kotlin/edu/usf/cutr/otp/plan/model/From.kt
@@ -1,9 +1,14 @@
-package edu.usf.cutr.otp.planner.model
+package edu.usf.cutr.otp.plan.model
 
 import kotlinx.serialization.Polymorphic
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Fare_ (
+data class From
+(
+    val name: String? = null,
+    val lon: Double? = null,
+    val lat: Double? = null,
+    val vertexType: String? = null,
     val additionalProperties: MutableMap<String,@Polymorphic Any> = HashMap()
 )
