+++ shared/src/commonMain/kotlin/edu/usf/cutr/otp/plan/model/From_.kt
@@ -1,14 +1,15 @@
-package edu.usf.cutr.otp.planner.model
+package edu.usf.cutr.otp.plan.model
 
 import kotlinx.serialization.Polymorphic
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class To_(
+data class From_
+(
     val name: String? = null,
     val lon: Double? = null,
     val lat: Double? = null,
-    val arrival: Long? = null,
+    val departure: Long? = null,
     val vertexType: String? = null,
     val additionalProperties: MutableMap<String,@Polymorphic Any> = HashMap()
 )
