+++ shared/src/commonMain/kotlin/edu/usf/cutr/otp/plan/model/LegGeometry.kt
@@ -1,10 +1,12 @@
-package edu.usf.cutr.otp.planner.model
+package edu.usf.cutr.otp.plan.model
 
 import kotlinx.serialization.Polymorphic
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Details
+data class LegGeometry
 (
+    val points: String? = null,
+    val length: Int? = null,
     val additionalProperties: MutableMap<String,@Polymorphic Any> = HashMap()
 )
