+++ shared/src/commonMain/kotlin/edu/usf/cutr/otp/plan/model/Metadata.kt
@@ -1,14 +1,14 @@
-package edu.usf.cutr.otp.planner.model
+package edu.usf.cutr.otp.plan.model
 
 import kotlinx.serialization.Polymorphic
 import kotlinx.serialization.Serializable
 
+
 @Serializable
-data class From
+data class Metadata
 (
-    val name: String? = null,
-    val lon: Double? = null,
-    val lat: Double? = null,
-    val vertexType: String? = null,
+    val searchWindowUsed: Int? = null,
+    val nextDateTime: Long? = null,
+    val prevDateTime: Long? = null,
     val additionalProperties: MutableMap<String,@Polymorphic Any> = HashMap()
 )
