+++ shared/src/commonMain/kotlin/edu/usf/cutr/otp/plan/model/Plan.kt
@@ -1,12 +1,15 @@
-package edu.usf.cutr.otp.planner.model
+package edu.usf.cutr.otp.plan.model
 
 import kotlinx.serialization.Polymorphic
 import kotlinx.serialization.Serializable
 
+
 @Serializable
-data class LegGeometry
+data class Plan
 (
-    val points: String? = null,
-    val length: Int? = null,
+    val date: Long? = null,
+    val from: From? = null,
+    val to: To? = null,
+    val itineraries: List<Itinerary>? = null,
     val additionalProperties: MutableMap<String,@Polymorphic Any> = HashMap()
 )
