+++ shared/src/commonMain/kotlin/edu/usf/cutr/otp/plan/model/To_.kt
@@ -1,15 +1,14 @@
-package edu.usf.cutr.otp.planner.model
+package edu.usf.cutr.otp.plan.model
 
 import kotlinx.serialization.Polymorphic
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class From_
-(
+data class To_(
     val name: String? = null,
     val lon: Double? = null,
     val lat: Double? = null,
-    val departure: Long? = null,
+    val arrival: Long? = null,
     val vertexType: String? = null,
     val additionalProperties: MutableMap<String,@Polymorphic Any> = HashMap()
 )
