--- shared/src/commonMain/kotlin/edu/usf/cutr/otp/planner/model/Plan.kt
@@ -1,15 +1,0 @@
-package edu.usf.cutr.otp.planner.model
-
-import kotlinx.serialization.Polymorphic
-import kotlinx.serialization.Serializable
-
-
-@Serializable
-data class Plan
-(
-    val date: Long? = null,
-    val from: From? = null,
-    val to: To? = null,
-    val itineraries: List<Itinerary>? = null,
-    val additionalProperties: MutableMap<String,@Polymorphic Any> = HashMap()
-)
